refactor(AuthInfo) : 패키지 분리로 인한 import 경로 설정
