+++ src/main/java/org/noostak/auth/application/AuthInfoServiceImpl.java
@@ -7,9 +7,9 @@ import org.noostak.auth.domain.AuthInfoRepository;
 import org.noostak.auth.domain.vo.AuthId;
 import org.noostak.auth.domain.vo.AuthType;
 import org.noostak.auth.domain.vo.RefreshToken;
-import org.noostak.auth.dto.AuthorizeResponse;
-import org.noostak.auth.dto.SignInResponse;
-import org.noostak.auth.dto.SignUpResponse;
+import org.noostak.auth.dto.common.AuthorizeResponse;
+import org.noostak.auth.dto.common.SignInResponse;
+import org.noostak.auth.dto.common.SignUpResponse;
 import org.noostak.member.domain.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
