+++ src/main/java/org/noostak/member/application/MemberServiceImpl.java
@@ -3,13 +3,11 @@ package org.noostak.member.application;
 import org.noostak.infra.KeyAndUrl;
 import org.noostak.infra.S3DirectoryPath;
 import org.noostak.infra.S3Service;
-import org.noostak.member.common.exception.MemberErrorCode;
-import org.noostak.member.common.exception.MemberException;
 import org.noostak.member.domain.Member;
 import org.noostak.member.domain.MemberRepository;
 import org.noostak.member.domain.vo.MemberName;
 import org.noostak.member.domain.vo.MemberProfileImageKey;
-import org.noostak.auth.dto.SignUpRequest;
+import org.noostak.auth.dto.common.SignUpRequest;
 import org.noostak.member.dto.GetProfileResponse;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
